[HOTFIX] fixed AnnotationException for Star Ratings system. Issue is that annoated class was never added to config file and HibernateUtils class

+++ database_scripts/create_longbox_db.sql
@@ -64,7 +64,7 @@ CREATE TABLE IF NOT EXISTS "star_ratings" (
     "id" bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     "comic_book_id" bigint,
     "user_id" bigint,
-    "rating" bigint,
+    "rating" integer
 );
 
 CREATE TABLE IF NOT EXISTS "comic_book_finished_list" (
